minor
Former-commit-id: 04c7e8b99b83f4640b610cc53277f01d85e57b84

+++ trip/iwave/core/lib/fd.cc
@@ -12,7 +12,7 @@ int fd_isarr(int i, IMODEL & model, IWaveInfo const & ic) {
   if (i<0 || i >= ic.get_num_fields()) {
     return 0;
   }
-  for (int j=0; j<model.active.size();j++) {
+  for (size_t j=0; j<model.active.size();j++) {
     if (ic.iwave_fields[i].field == model.active[j]) {
       return 1;
     }
